ai-dev-team-moe-v4_extracted/ai-dev-team-moe-v4/requirements.txt

streamlit>=1.37.0
pyyaml>=6.0.1
pydantic>=2.8.2
tenacity>=8.2.3
